Silence warnings

+++ test/class/opal_fifo.c
@@ -48,7 +48,7 @@ static void *thread_test (void *arg) {
     timing = ((double) total.tv_sec + (double) total.tv_usec * 1e-6) / (double) ITERATIONS;
 
     printf ("Atomics thread finished. Time: %d s %d us %d nsec/poppush\n", (int) total.tv_sec,
-            total.tv_usec, (int)(timing / 1e-9));
+            (int)total.tv_usec, (int)(timing / 1e-9));
 
     return NULL;
 }
@@ -84,7 +84,7 @@ static void *thread_test_exhaust (void *arg) {
   timing = ((double) total.tv_sec + (double) total.tv_usec * 1e-6) / (double) item_count;
 
   fprintf (stderr, "Exhaustive atomics thread finished. Popped %d items. Time: %d s %d us %d nsec/poppush\n", item_count,
-	   (int) total.tv_sec, total.tv_usec, (int)(timing / 1e-9));
+	   (int) total.tv_sec, (int)total.tv_usec, (int)(timing / 1e-9));
 
   return NULL;
 }
@@ -167,7 +167,7 @@ int main (int argc, char *argv[]) {
     }
 
     printf ("Single thread test. Time: %d s %d us %d nsec/poppush\n", (int) total.tv_sec,
-            total.tv_usec, (int)(timing / 1e-9));
+            (int)total.tv_usec, (int)(timing / 1e-9));
 
     thread_test (&fifo);
 
@@ -200,7 +200,7 @@ int main (int argc, char *argv[]) {
     }
 
     printf ("All threads finished. Thread count: %d Time: %d s %d us %d nsec/poppush\n",
-             OPAL_FIFO_TEST_THREAD_COUNT, (int) total.tv_sec, total.tv_usec, (int)(timing / 1e-9));
+            OPAL_FIFO_TEST_THREAD_COUNT, (int) total.tv_sec, (int)total.tv_usec, (int)(timing / 1e-9));
 
 
     gettimeofday (&start, NULL);
@@ -226,7 +226,7 @@ int main (int argc, char *argv[]) {
     }
 
     printf ("All threads finished. Thread count: %d Time: %d s %d us %d nsec/poppush\n",
-             OPAL_FIFO_TEST_THREAD_COUNT, (int) total.tv_sec, total.tv_usec, (int)(timing / 1e-9));
+            OPAL_FIFO_TEST_THREAD_COUNT, (int) total.tv_sec, (int)total.tv_usec, (int)(timing / 1e-9));
 
     success = true;
     for (int i = 0 ; i < ITEM_COUNT ; ++i) {
